+++ tests/unit/test_exchange_rates.py
@@ -15,17 +15,6 @@ def test_value_after_connetion_to_json():
     assert type(source.to_json()) is dict
 
 
-"""
 def test_value_after_connetion_to_df():
     source = ExchangeRates()
     assert type(source.to_df()) is pd.DataFrame
-
-
-
-def test_value_after_connetion_to_df():
-    source = ExchangeRates()
-    assert source.to_df().isna().any()
-"""
-
-# def test_to_df():
-#     pass